Question3
You are required to compute the power of a number by implementing a calculator. Create a class MyCalculator which consists of a single method “long power(int, int)”. This method takes two integers, n  and p, as parameters and finds p times n. If either n or p  is negative, then the method must throw an exception which says "n or p should not be negative". Also, if both n and p  are zero, then the method must throw an exception which says "n and p should not be zero".
For example, -4 and -5 would result in. java.lang.Exception n or p should not be negative
Implement the class MyCalculator and write the function “power”  return the appropriate result after the power operation or an appropriate exception as detailed above.