Java File I/O Question 5
Write a java program to read a CSV file, identify the data segregated in different columns, and print the data onto a .txt file, however, the data may be present in the CSV file in any order (i.e. not necessarily the First Name at index 0, Last Name at index 1, etc ). Hint: Use the concept of collections in java.